Assembly: rename the pattern to ScatterMap, move the kernels into the forms
NamedFormPattern was a misnomer: the sparsity pattern proper lives in the
assembled Math::SparseMatrix, and what the class stores is the destination
of each cell-local entry in that matrix's value array. Renames it to
ScatterMap, which says the direction without colliding with "local to
global map", which already names the DOF map in every finite element
vocabulary and is getDOFs(d, i) here.

The cell kernels are mathematical objects belonging to the forms they
integrate, not to the assembly backends, so MassFormKernel and
DiffusionFormKernel move from Assembly/NamedFormKernel.h into MassForm.h
and DiffusionForm.h. Each form now names its own kernel through
KernelType, and the assembly specializations ask the form for it instead
of hardcoding one, which is also what lets a new named form reuse the
same two specializations.

No behavioral change; the benchmarks are unchanged within noise.
