+++ src/Rodin/Assembly/CMakeLists.txt
@@ -7,8 +7,7 @@ Distributed under the Boost Software License, Version 1.0.
 
 set(RodinAssembly_HEADERS
   AssemblyBase.h
-  NamedFormKernel.h
-  NamedFormPattern.h
+  ScatterMap.h
   Sequential.h
 )
