+++ src/Rodin/Assembly/Sequential.h
@@ -35,8 +35,7 @@
 #include "Rodin/Assembly/ConstraintMap.h"
 
 #include "ForwardDecls.h"
-#include "NamedFormKernel.h"
-#include "NamedFormPattern.h"
+#include "ScatterMap.h"
 
 namespace Rodin::Assembly
 {
@@ -124,7 +123,7 @@ namespace Rodin::Assembly
         SequentialIteration seq(mesh, Geometry::Region::Cells);
         const size_t d = seq.getDimension();
         const Index count = seq.getCount();
-        m_pattern.template assemble<MassFormKernel<Scalar, TrialFES, TestFES>>(
+        m_scatterMap.template assemble<typename InputType::KernelType>(
           out, trialFES, testFES, seq, d, count);
       }
 
@@ -134,7 +133,7 @@ namespace Rodin::Assembly
       }
 
     private:
-      mutable NamedFormPattern<Scalar> m_pattern;
+      mutable ScatterMap<Scalar> m_scatterMap;
   };
 
   template <class Solution, class TrialFES, class TestFES, class Scalar>
@@ -162,7 +161,7 @@ namespace Rodin::Assembly
         SequentialIteration seq(mesh, Geometry::Region::Cells);
         const size_t d = seq.getDimension();
         const Index count = seq.getCount();
-        m_pattern.template assemble<DiffusionFormKernel<Scalar, TrialFES, TestFES>>(
+        m_scatterMap.template assemble<typename InputType::KernelType>(
           out, trialFES, testFES, seq, d, count);
       }
 
@@ -172,7 +171,7 @@ namespace Rodin::Assembly
       }
 
     private:
-      mutable NamedFormPattern<Scalar> m_pattern;
+      mutable ScatterMap<Scalar> m_scatterMap;
   };
 
   /**
